requirements.txt
openai==0.28.1
selenium==4.10.0
beautifulsoup4==4.12.2
requests==2.31.0
newspaper3k==0.2.8
pydantic==1.10.13
psycopg2-binary==2.9.9
Flask==2.2.5
lxml==4.9.3
python-dotenv==1.0.0
jinja2==3.1.2
gunicorn==21.2.0
python-dateutil==2.8.2
Werkzeug==2.2.3
itsdangerous==2.1.2
MarkupSafe==2.1.2
click==8.1.7

